+++ packages/components-js/projects/partials/scripts/generateBrowserSupportFallbackScriptPartial.ts
@@ -5,9 +5,7 @@ export const generateBrowserSupportFallbackScriptPartial = (): string => {
   const types = `type GetBrowserSupportFallbackScriptOptions = {
   cdn?: Cdn;
   format?: Format;
-};
-type GetBrowserSupportFallbackScriptOptionsFormatHtml = GetBrowserSupportFallbackScriptOptions & { format: 'html' };
-type GetBrowserSupportFallbackScriptOptionsFormatJsx = GetBrowserSupportFallbackScriptOptions & { format: 'jsx' };`;
+};`;
 
   const fallbacksFilePath = require.resolve('@porsche-design-system/fallbacks');
   const packageDir = path.resolve(path.dirname(fallbacksFilePath), '../..');
@@ -17,8 +15,8 @@ type GetBrowserSupportFallbackScriptOptionsFormatJsx = GetBrowserSupportFallback
     .replace('https://cdn.ui.porsche.com', '${cdnBaseUrl}')
     .trim();
 
-  const func = `export function getBrowserSupportFallbackScript(opts?: GetBrowserSupportFallbackScriptOptionsFormatHtml): string;
-export function getBrowserSupportFallbackScript(opts?: GetBrowserSupportFallbackScriptOptionsFormatJsx): JSX.Element;
+  const func = `export function getBrowserSupportFallbackScript(opts?: GetBrowserSupportFallbackScriptOptions & { format: 'jsx' }): JSX.Element;
+export function getBrowserSupportFallbackScript(opts?: GetBrowserSupportFallbackScriptOptions): string;
 export function getBrowserSupportFallbackScript(opts?: GetBrowserSupportFallbackScriptOptions): string | JSX.Element {
   const { cdn, format }: GetBrowserSupportFallbackScriptOptions = {
     cdn: 'auto',
@@ -33,7 +31,9 @@ export function getBrowserSupportFallbackScript(opts?: GetBrowserSupportFallback
 
   // there is no other solution than using dangerouslySetInnerHTML since JSX elements are rendered by the createElement() function
   // https://stackoverflow.com/a/64815699
-  return format === 'html' ? \`<script>\${scriptContent}</script>\` : <script dangerouslySetInnerHTML={{__html: scriptContent}} />;
+  return format === 'html'
+    ? \`<script>\${scriptContent}</script>\`
+    : <script dangerouslySetInnerHTML={{ __html: scriptContent }} />;
 }`;
 
   return [types, func].join('\n\n');