+++ packages/components-js/tests/smoke/specs/cdn.smoke.ts
@@ -10,6 +10,8 @@ import {
   MARQUES_MANIFEST,
   META_ICONS_CDN_BASE_URL,
   META_ICONS_MANIFEST,
+  MODEL_SIGNATURES_MANIFEST,
+  MODEL_SIGNATURES_CDN_BASE_URL,
 } from '@porsche-design-system/assets';
 import { getFontFaceStylesheet } from '@porsche-design-system/components-js/partials';
 import { COMPONENT_CHUNKS_MANIFEST } from '../../../projects/components-wrapper/lib/chunksManifest';
@@ -158,7 +160,7 @@ describe('cdn', () => {
         })(item);
       }
 
-      it(`should have all ${items.length} ${baseUrl.substr(baseUrl.lastIndexOf('/') + 1)}`, () => {
+      it(`should have all ${items.length} ${baseUrl.substring(baseUrl.lastIndexOf('/') + 1)}`, () => {
         expect(responseCounter).toBe(items.length);
         responseCounter = 0; // reset for upcoming test
       });
@@ -195,15 +197,23 @@ describe('cdn', () => {
       bulkRequestItems(metaIcons, META_ICONS_CDN_BASE_URL);
     });
 
+    describe('model-signatures', () => {
+      const modelSignatures = objectToFlatArray(MODEL_SIGNATURES_MANIFEST);
+      bulkRequestItems(modelSignatures, MODEL_SIGNATURES_CDN_BASE_URL);
+    });
+
     describe('styles', () => {
       // retrieve css file names via partial since FONT_FACE_CDN_URL returns different value based on flag in window
-      const comStyle = getFontFaceStylesheet({ cdn: 'auto', withoutTags: true });
-      const cnStyle = getFontFaceStylesheet({ cdn: 'cn', withoutTags: true });
+      const comStyle = getFontFaceStylesheet({ cdn: 'auto' });
+      const cnStyle = getFontFaceStylesheet({ cdn: 'cn' });
 
       // extract file name from full path
-      const getFileName = (path: string) => path.substr(path.lastIndexOf('/') + 1);
+      const getFileName = (path: string): string => {
+        path = path.replace(/(.*)(https:\/\/[a-z0-9./-]+\.css)(.*)/, '$2');
+        return path.substring(path.lastIndexOf('/') + 1);
+      };
 
-      const styles = [getFileName(comStyle), getFileName(cnStyle)];
+      const styles = [comStyle, cnStyle].map(getFileName);
       const baseUrl = `${CDN_BASE_URL}/${CDN_BASE_PATH_STYLES}`;
       bulkRequestItems(styles, baseUrl);
     });