+++ packages/components-js/tests/vrt/puppeteer/specs/component-scaling.vrt.ts
@@ -8,10 +8,12 @@ const components = [
   'accordion',
   'banner',
   'button',
+  'button-group',
   'button-pure',
   'checkbox-wrapper',
   'content-wrapper',
   'display',
+  'divider',
   'fieldset-wrapper',
   'heading',
   'headline',
@@ -35,6 +37,7 @@ const components = [
   'tag-dismissible',
   'text-field-wrapper',
   'textarea-wrapper',
+  'text-list',
   'toast-basic',
 ];