+++ packages/components/src/components/text-list/text-list-item/text-list-item-styles.spec.ts
@@ -1,12 +1,7 @@
 import { getComponentCss } from './text-list-item-styles';
 
-xdescribe('getComponentCss()', () => {
-  it.each<Parameters<typeof getComponentCss>>([
-    ['unordered', 'numbered'],
-    ['unordered', 'alphabetically'],
-    ['ordered', 'numbered'],
-    ['ordered', 'alphabetically'],
-  ])('should return correct css for listType: %s and orderType: %s', (...args) => {
+describe('getComponentCss()', () => {
+  it.each<Parameters<typeof getComponentCss>>([[]])('should return correct css', (...args) => {
     expect(getComponentCss(...args)).toMatchSnapshot();
   });
 });