+++ packages/storefront/storefront.config.ts
@@ -261,6 +261,14 @@ export const config: StorefrontConfig = {
       // @ts-ignore
       Props: [() => import('@/../../components/src/components/modal/modal.props.md')],
     },
+    'Model Signature': {
+      // @ts-ignore
+      Examples: [() => import('@/../../components/src/components/model-signature/model-signature.examples.md')],
+      // @ts-ignore
+      Usage: [() => import('@/../../components/src/components/model-signature/model-signature.usage.md')],
+      // @ts-ignore
+      Props: [() => import('@/../../components/src/components/model-signature/model-signature.props.md')],
+    },
     Notifications: {
       // @ts-ignore
       Introduction: [() => import('@/pages/components/notifications/notifications.usage.md')],
@@ -269,7 +277,7 @@ export const config: StorefrontConfig = {
         () => import('@/../../components/src/components/inline-notification/inline-notification.examples.md'),
       ],
       // @ts-ignore
-      Toast: [() => import('@/../../components/src/components/toast/toast/toast.examples.md')],
+      Toast: [() => import('@/../../components/src/components/toast/toast.examples.md')],
       // @ts-ignore
       Banner: [() => import('@/../../components/src/components/banner/banner.examples.md')],
       // @ts-ignore