+++ packages/components/CHANGELOG.md
@@ -12,6 +12,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 #### Changed
 
 - Background color of `Scroller`'s `prev` and `next` buttons in dark theme
+- Partials: Removed deprecated `withoutTags` option for all partials, please use `format: 'jsx'` instead
+- `Divider` and `Button Group` match new design language.
+- `Text List` matches new design language
+
+#### Added
+
+- `Model Signature`
 
 ### [3.0.0-alpha.1] - 2023-02-16
 
@@ -23,7 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Changed
 
-- `Stepper Horizontal` matches new design language.
+- `Stepper Horizontal` matches new design language
 - Styles: Optimize design tokens "spacing", "typography" and "theme" provided by styles sub-package
   `@porsche-design-system/components-{js|angular|react|vue}/styles`
 - Styles: Use calc() instead of max() to calculate padding for `gridStyle` (JS) and `pds-grid` (SCSS)
@@ -409,6 +416,21 @@ to the new values since those ones will be removed with next major version.
 + …addMessage({ text: `Some message`, state: 'info' })
 ```
 
+##### Divider:
+
+- Prop values `neutral-contrast-low | neutral-contrast-medium | neutral-contrast-high` of `color` prop are deprecated.
+
+```diff
+- <p-divider color="neutral-contrast-low"></p-divider>
++ <p-divider color="contrast-low"></p-divider>
+
+- <p-divider color="neutral-contrast-medium"></p-divider>
++ <p-divider color="contrast-medium"></p-divider>
+
+- <p-divider color="neutral-contrast-high"></p-divider>
++ <p-divider color="contrast-high"></p-divider>
+```
+
 #### Added
 
 - `Display` component
@@ -425,10 +447,10 @@ to the new values since those ones will be removed with next major version.
   `Radio Button Wrapper`, `Popover`, `Modal`, `Select Wrapper`, `Tag`, `Tag Dismissible`, `Textarea Wrapper`,
   `Inline Notification`, `Banner`, `Toast`, `Grid`, `Flex`, `Pagination`, `Scroller`, `Accordion`, `Text`,
   `Text Field Wrapper`, `Content Wrapper`, `Segmented Control`, `Tabs`, `Tabs Bar`, `Headline` and `Fieldset Wrapper`
-  matches new design language.
+  match new design language
 - `Icon` supports
   `primary | contrast-low | contrast-medium | contrast-high | notification-success | notification-warning | notification-error | notification-info | inherit`
-  for `color` prop.
+  for `color` prop
 - Default value of prop `width` of `Banner` has changed from `basic` to `extended`
 - Default value of prop `action-icon` of `Inline Notification` has changed from `arrow-head-right` to `arrow-right`
 - Default value of prop `name` of `Icon` has changed from `arrow-head-right` to `arrow-right`